# Add a better injectModule method to WASM