--- a/vm/src/obj/objframe.rs
+++ b/vm/src/obj/objframe.rs
@@ -31,7 +31,7 @@ impl FrameRef {
 
     #[pyproperty]
     fn f_globals(self, vm: &VirtualMachine) -> PyResult<PyDictRef> {
-        if self.code.incognito {
+        if self.code.incognito() {
             Err(vm.new_type_error("Can't get f_globals on an incognito frame".to_owned()))
         } else {
             Ok(self.scope.globals.clone())
@@ -40,7 +40,7 @@ impl FrameRef {
 
     #[pyproperty]
     fn f_locals(self, vm: &VirtualMachine) -> PyResult<PyDictRef> {
-        if self.code.incognito {
+        if self.code.incognito() {
             Err(vm.new_type_error("Can't get f_locals on an incognito frame".to_owned()))
         } else {
             Ok(self.scope.get_locals())