--- a/vm/src/obj/objfunction.rs
+++ b/vm/src/obj/objfunction.rs
@@ -280,12 +280,8 @@ impl PyFunction {
     }
 
     #[pyproperty(magic)]
-    fn globals(&self, vm: &VirtualMachine) -> PyResult<PyDictRef> {
-        if self.code.incognito() {
-            Err(vm.new_type_error("Can't get __globals__ on an incognito function".to_owned()))
-        } else {
-            Ok(self.scope.globals.clone())
-        }
+    fn globals(&self) -> PyDictRef {
+        self.scope.globals.clone()
     }
 }