--- a/vm/src/obj/objfunction.rs
+++ b/vm/src/obj/objfunction.rs
@@ -278,6 +278,15 @@ impl PyFunction {
     fn kwdefaults(&self) -> Option<PyDictRef> {
         self.kw_only_defaults.clone()
     }
+
+    #[pyproperty(magic)]
+    fn globals(&self, vm: &VirtualMachine) -> PyResult<PyDictRef> {
+        if self.code.incognito {
+            Err(vm.new_type_error("Can't get __globals__ on an incognito function".to_owned()))
+        } else {
+            Ok(self.scope.globals.clone())
+        }
+    }
 }
 
 #[pyclass]