… benchmarks/bench.rs
… examples/dis.rs
… tests/snippets/invalid_syntax.py
… wasm/tests/test_demo.py
… wasm/tests/test_inject_module.py