This allows for strict privacy for WASM modules, which is important in an environment like the browser, where most of the time you're running code on someone else's computer. The idea for this method would be to write a wrapper module around some possibly security-sensitive functions that you pass from js, and then allow arbitrary code to access that module. For [robot rumble](https://github.com/chicode/robot-rumble) (the successor to codingworkshops, the code learning platform), it might be used like:

```js
// vm.injectModule(name, source, imports?, strict_private?)
vm.injectModule(
    "robots",
    `
__all__ = ["export"]
def export(...): runRobot(...)
# robot-rumble "stdlib" source code
`,
    { runRobot: () => { ... } },
    true,
);
// later
vm.exec(userCode);
```

We don't want user code arbitrarily calling the `runRobot` function, so we wrap it in a module that makes sure that it *can't*, and also (ideally) provides a nicer, more pythonic interface. The `strict_private` parameter enables that; it runs the source code provided in one namespace with the JS imports, and copies only the items exported in `__all__` to a new namespace that is injected into `sys.modules`.