@coolreader18 2020-02-04
Hmm, just realized that it would be possible to access the globals by doing e.g. `get_robot.__globals__['runRobot']`. I might look into preventing that, though I don't know if it's that necessary; it's pretty obscure.