@coolreader18 2020-02-05
Alright, I added the concept of "incognito" mode (though it's not an actual `compile::Mode`), which means that functions defined in it cannot have their `__globals__` accessed, and frames that originate from an incognito context disallow accessing `f_globals` and `f_locals`. There's probably a few things like that that I missed, but I think the basic concept is decent.