@youknowone 2020-02-07
Because I didn't look into non-vm part of this project that much, I am confused about this a little. To make it clear, this is a extension feature (mostly) for wasm, right?

Basically, I don't like the idea that non-standard python feature slow down the standard features.
It seems this feature is only enabled when `inject_modlue` is called. So please review if this idea will work:
1. Define new objects SecureModule, SecureFrame and SecureCodeObject (or Incognito*)
2. Define traits of module, frame and code object. Because the mode is given to compiler, a single module will have only CodeObject or SecureCodeObject. So they probably perfectly exclude each other.
3. Turn some functions and struct into generic. Probably vm.run_code, PyFunction and module related things?
4. These things might be placed in wasm or vm by whether this feature is belong to rustpython or rustpython-wasm.

Note: Here "standard Python" means documented or CPython compatible.