@coolreader18 2020-02-07
I think that _would_ be a better way of separating the concerns, but it might also be cumbersome to have to define `CodeObject`, `Frame`, and `PyFunction` twice. Making core functions/frame code generic might also increase the code size a _ton_. I think this is sort of about whether or not we should implement features that CPython doesn't have, so here's my take on that: as an alternative Python interpreter, we are in a unique situation where we can experiment with non-standard features that we consider useful, and see how they work/interact with the rest of the language, without having to go through a PEP process first. At least in this situation (e.g. arbitrary Python code running in the browser), a feature like this is definitely necessary, but CPython doesn't usually run in the browser. However, I'm sure there are other environments where Python code does need security like the security that this feature provides, so I think it would be interesting to try this and see where else it could be used.