@coolreader18 2020-02-07
Also, w.r.t. "only used from the `rustpython-wasm` module", I was thinking that it could also be enabled from a Python module by something like `from __future__ import incognito_module` or `__incognito__ = true`.