@palaviv 2020-02-08
> Basically, I don't like the idea that non-standard python feature slow down the standard features.

@coolreader18 can you give us some numbers on the performance impact this change has?