@coolreader18 2020-02-09
Alright, I had some trouble getting the benchmarks to work, but here's what I got:

On master:
```
running 3 tests
test bench_rustpy_mandelbrot    ... bench: 1,017,088,271 ns/iter (+/- 21,787,720)
test bench_rustpy_nbody         ... bench: 306,066,689 ns/iter (+/- 13,873,395)
test bench_rustpy_parse_to_ast  ... bench:  10,004,214 ns/iter (+/- 1,026,285) = 6 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 3 measured; 4 filtered out


RustPython on  master [!] is 📦 v0.1.1 via 🐍 v3.8.1 via 🦀 v1.41.0 took 6m42s 
❯ 
```
On this branch:
```

running 3 tests
test bench_rustpy_mandelbrot    ... bench: 1,028,623,787 ns/iter (+/- 51,170,076)
test bench_rustpy_nbody         ... bench: 305,318,476 ns/iter (+/- 30,826,374)
test bench_rustpy_parse_to_ast  ... bench:  10,042,795 ns/iter (+/- 689,063) = 6 MB/s

test result: ok. 0 passed; 0 failed; 0 ignored; 3 measured; 4 filtered out


RustPython on  coolreader18/wasm-inject-module [!] is 📦 v0.1.1 via 🐍 v3.8.1 via 🦀 v1.41.0 took 6m50s 
❯ 
```

Basically very little difference; the difference between 2 benchmarks are within both the standard deviations given by the bencher.