@palaviv 2020-02-09
I agree that this is a very small decrease in performance. What do you think @youknowone?
We should document features that are not part of CPython. Where do you think this should be documented? I mentioned it in #1192 but we never decided where.