@youknowone 2020-02-11
To make things clear, I am not asking changes or trying to decline this PR by these comments. This is just my general preferences.

Before talking about performance, the performance itself doesn't justify the design. So this is a kind of negotiation of trade-off between performance and a new feature. It seems having multiple `PyFunction` is not a big deal. If generic has too much complexity for core python features, then still secure objects can take over the complexity part as an optional feature. I think `struct SecureCodeObject(CodeObject)` and similar secure frame is another option. it seems we don't need 2 versions of frame - we just need 2 versions of FrameRef.

Being included in rustpython core or wasm is subtle issue. How about this? wasm support is one of the major feature of RustPython but perfectly living out of rustpython core world. So no matter whatever things happen in there. I think any non-standard feature can be added in this way of extension. We just can keep the core for the best compatibility.

About the benchmark, the measured performance doesn't look meaningful. The benchmark should measure the cost of `f_globals` and `f_locals`. The benchmark code use them enough? I am doubtful in general sense of Python though I didn't look in the benchmark test code.

Still, if you guys think adding this feature in this way is better than other way, I think it is acceptable. Because usually I was concentrating on core features, I don't have any perspective of the big plans of the project. So I don't want to disturb to decide things we do out of the core language.