@coolreader18 2020-02-11
That makes a lot of sense; I think your idea of separate `FrameRef` payoad types is much more elegant than my solution, I just don't know if it's feasible with how the `PyRef` type works at the moment. Eventually I'd like to allow some sort of `upcast` operation on a PyObject that returns a reference to the payload type of its base class, but I'm still trying to figure out the best way to do that. That would be necessary to implement your solution, though, as `FrameRef`s stored in e.g. the `vm.frames` frame stack are just `PyRef`s and can be accessed from Python, so it wouldn't be possible to store multiple different Frame payloads in the frame stack without that upcast feature.