@windelbouwman 2020-02-21
Hi all,

I looked into this change a bit, and I cannot say I agree with it. Python is insecure by design, you can modify all kind of things. In my opinion, adding an incognito mode adds extra code which tries to protect global variables, but as you mentioned elsewhere, you could still circumvent this by accessing the globals in another obscure way. I think it would be best to stick to CPython behavior, which is insecure by design. Making python properly secure can probably be harder than you imagine at first.

I like the `__future__` compilation part of the change, that would be good to merge, but I'm reluctant to merge the incognito mode. Also, I would like to be able to use the rustpython code as a library on embedded systems, and then I want the library to be as simple as possible, and not too tailered towards WASM / javascript.

Please do not be offended by this feedback, but this is my view on the project.